% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_hw_doc.R
\docType{data}
\name{dat_hw}
\alias{dat_hw}
\title{\code{dat_hw} dataset}
\format{
A tibble with 100 rows and 2 variables
\describe{
\item{weight}{a number denoting weight in TODO}
\item{height}{a number denoting height in TODO}
}
}
\usage{
dat_hw
}
\description{
The \code{dat_hw} sample dataset used in The Missing Data Book exercises
}
\keyword{datasets}
