% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frogger_doc.R
\docType{data}
\name{frogger}
\alias{frogger}
\title{\code{frogger} dataset}
\format{
A tibble with 15 rows and 4 variables
\describe{
\item{name}{a character denoting the player name}
\item{time}{a character denoting the time of day}
\item{value}{a number denoting the frogger score}
\item{year}{a number denoting the year}
}
}
\usage{
frogger
}
\description{
The \code{frogger} sample dataset used in The Missing Data Book exercises TODO check descriptions
}
\keyword{datasets}
