% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacman_doc.R
\docType{data}
\name{pacman}
\alias{pacman}
\title{\code{pacman} dataset}
\format{
A tibble with 2000 rows and 6 variables
\describe{
\item{year}{a number denoting the year}
\item{month}{a number denoting the month}
\item{day}{a number denoting the day of the month}
\item{initial}{a character denoting players' initials}
\item{score}{a character containing player scores, and missing values as "N/A"}
\item{country}{a character containing the two-letter abbreviation for player country}
}
}
\source{
{TODO Add source for pacman data}
}
\usage{
pacman
}
\description{
The \code{pacman} sample dataset used in The Missing Data Book exercises, containing pacman game scores.
}
\keyword{datasets}
