% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_doc.R
\docType{data}
\name{pie}
\alias{pie}
\title{\code{pie} dataset}
\format{
A tibble with 8 rows and 3 variables
\describe{
\item{type}{a character denoting the type of pie}
\item{bakery}{a character denoting the bakery name}
\item{price}{a number denoting the price per pie, in dollars}
}
}
\usage{
pie
}
\description{
The \code{pie} sample dataset used in The Missing Data Book exercises
}
\keyword{datasets}
